(cloudwatch) divide metric find query
